Initial commit of model files.
The way this project is built will change a little bit now that it is in
its own repo.

src is for the raw source
dist is for the compiled usable files
test is for tests!

The request files will most likely be upgraded to ES6 so that all the
files can be compiled through traceur to multiple output types (amd,
commonjs, es6 imports, etc);

Once ready this repo will be added as a git submodule to the web app repo.