Internal Updates, now using Symbols.
Started using Symbols to do some element hiding.
NOTE: Next version of Traceur will use native Symbols over their Symbol
shim. Will update as soon as its released.

'social' property of MHObjects is now a get/set combination. The
MHSocial object is stored on the instance via a Symbol. The social set
function does not allow setting the internal symbol property to anything
but an instanceof MHSocial or null. If the internal property is set to
undefined social will still return null.
'fetchSocial' will now fetch if force === true or 'social' is not an
MHSocial instance.

MHCache's internal keymap is now stored as a Symbol property

MHObject.create is now the only place objects are stored into the
MHCache instance. MHImages are not cached, but Actions are.

MHRelationalPair now checks the MHCache instance of the 'object' and
returns the Cache'd object before creating a new instance.